requirements.txt
discord.py
python-dotenv
yt-dlp
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
ffmpeg-python
